Day 1: Add first Python programs

+++ Python/Day1/currency_converter.py
@@ -1,4 +1,4 @@
 exchange_rate = 363.14
 dollars = float(input("How many dollars do you have? "))
-drams = dollars * exchange_rate
+drams = int(dollars * exchange_rate)
 print (f"You have {drams} drams!")
